# Reduce onenormest memory usage