@argriffing 2015-04-13
Thanks!  I think this PR together with the merged PR https://github.com/scipy/scipy/pull/4546 will have covered all of the modifications in https://github.com/scipy/scipy/pull/4095.

I think the slowdown for small matrices is OK for the reasons you've given.

One nitpick is that this PR uses `scipy.sparse.rand` which is fine in many contexts but which is a particularly weak choice for testing or benchmarking onenormest because `rand` samples only non-negative values.  A feature of the onenormest implementation is that it computes one-norms of entrywise non-negative matrices 'exactly' (with the usual floating point error, but without the more interesting sources of error), basically using the following strategy:

```python
>>> A = np.random.rand(20, 20)
>>> np.linalg.norm(A, ord=1)
12.716998703114299
>>> A.T.dot(np.ones(20)).max()
12.716998703114299
```

It does this implicitly without having to look at the entries of the matrix.  The cleverness of the algorithm does not come into play for these matrices, so tests or benchmarks that use them could be improved by also using linear operators that have entrywise mixed signs.