@argriffing 2015-04-16
Merged, thanks @pv!